+++ setup.sh
@@ -22,3 +22,15 @@ then
   mkdir shared && cd "$_"
   mkdir web
 fi
+
+if [ ! -f "./deploy.log" ]
+then
+  printf -- "Create empty log file. \n"
+  touch deploy.log
+fi
+
+if [ ! -f "./.env.example" ]
+then
+  printf -- "Rename .env.example. \n"
+  mv .env.example .env
+fi